+++ triangulo.py
@@ -1,8 +1,8 @@
 #Area de um triangulo#
 
 def area_triangulo():
-    b = int(input("Digte o valor de B: "))
-    h = int(input("Digite o valor de H: "))
+    b = float(input("Digte o valor de B: "))
+    h = float(input("Digite o valor de H: "))
     
     a = (b * h) / 2
     print(f'O resultado da área do triangulo é {a}')
